requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-multipart
python-docx
PyPDF2
spacy
nltk
scikit-learn
pydantic
python-dotenv